Added cargo file in sql utils
