+++ crates/proof-of-sql-utils/Cargo.toml
@@ -1,0 +1,39 @@
+[package]
+name = "proof_of_sql_utils"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+ahash = { workspace = true }
+ark-bls12-381 = { workspace = true }
+ark-curve25519 = { workspace = true }
+ark-ec = { workspace = true }
+ark-ff = { workspace = true }
+ark-poly = { workspace = true }
+ark-serialize = { workspace = true }
+ark-std = { workspace = true }
+arrow = { workspace = true, optional = true }
+bit-iter = { workspace = true }
+bigdecimal = { workspace = true }
+blake3 = { workspace = true }
+blitzar = { workspace = true, optional = true }
+bumpalo = { workspace = true, features = ["collections"] }
+bytemuck = { workspace = true }
+byte-slice-cast = { workspace = true }
+curve25519-dalek = { workspace = true, features = ["serde"] }
+chrono = { workspace = true, features = ["serde"] }
+derive_more = { workspace = true }
+indexmap = { workspace = true, features = ["serde"] }
+itertools = { workspace = true }
+merlin = { workspace = true, optional = true }
+num-traits = { workspace = true }
+num-bigint = { workspace = true, default-features = false }
+postcard = { workspace = true, features = ["alloc"] }
+rand = { workspace = true, default-features = false, optional = true }
+rayon = { workspace = true, optional = true }
+serde = { workspace = true, features = ["serde_derive"] }
+serde_json = { workspace = true }
+snafu = { workspace = true }
+tiny-keccak = { workspace = true }
+tracing = { workspace = true, features = ["attributes"] }
+zerocopy = { workspace = true }
